# docs/specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: chatlab HTTP API
  version: 1.0.0-rc.1
  description: |
    chatlab is a local development platform for chat agents — workspaces, multi-provider LLMs,
    feedback corpus.

    All endpoints under `/v1/*` require `Authorization: Bearer <token>`. Token is permissive by
    default; set `CHATLAB_REQUIRE_TOKEN=<value>` for strict mode. `/healthz`, `/readyz`,
    `/favicon.ico` are unauthenticated.

    Errors are returned as `{ error: { message, type, code, error_subcode?, error_data?, request_id } }`.
  license:
    name: MIT
    url: https://github.com/jvrmaia/chatlab/blob/main/LICENSE
  contact:
    url: https://github.com/jvrmaia/chatlab
servers:
  - url: http://127.0.0.1:4480
    description: Default local chatlab
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
...
    AgentId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: Agent UUID.
...
    MediaId:
      in: path
      name: mediaId
      required: true
      schema: { type: string, format: uuid }
      description: Media UUID.
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: Upstream provider error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: Upload exceeds the size cap
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnsupportedMediaType:
      description: Mime type doesn't match declared media type
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Removed:
      description: Resource removed
      content:
        application/json:
          schema:
            type: object
            required: [removed]
            properties:
              removed: { type: boolean, enum: [true] }
  schemas:
...
    Agent:
      type: object
      required: [id, workspace_id, name, provider, model, context_window, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        name: { type: string }
        provider:
          type: string
          enum: [openai, anthropic, deepseek, gemini, maritaca, ollama, custom]
        model: { type: string }
        api_key:
          type: string
          description: "Masked in HTTP responses (`***last4`)."
        base_url: { type: string }
        system_prompt: { type: string }
        context_window: { type: integer, minimum: 1, maximum: 200, default: 20 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AgentCreate:
      type: object
      required: [name, provider, model]
      properties:
        name: { type: string }
        provider:
          type: string
          enum: [openai, anthropic, deepseek, gemini, maritaca, ollama, custom]
        model: { type: string }
        api_key: { type: string }
        base_url: { type: string }
        system_prompt: { type: string }
        context_window: { type: integer, minimum: 1, maximum: 200, default: 20 }
    AgentPatch:
      type: object
      properties:
        name: { type: string }
        provider:
          type: string
          enum: [openai, anthropic, deepseek, gemini, maritaca, ollama, custom]
        model: { type: string }
        api_key:
          type: string
          description: "Omit (or send empty string) to preserve the stored key."
        base_url: { type: string }
        system_prompt: { type: string }
        context_window: { type: integer, minimum: 1, maximum: 200 }
...
    MediaRecord:
      type: object
      required: [id, type, mime_type, size, sha256, created_at]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [image, audio, video, document, sticker]
        mime_type: { type: string }
        size: { type: integer }
        sha256: { type: string }
        filename: { type: string }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code, request_id]
          properties:
            message: { type: string }
            type: { type: string, description: "Always `ApiError`." }
            code: { type: integer }
            error_subcode:
              oneOf:
                - { type: string }
                - { type: integer }
            error_data:
              type: object
              additionalProperties: true
            request_id: { type: string }
security:
  - bearer: []
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: List agents in the active workspace (api_keys masked)
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Agents]
      operationId: createAgent
      summary: Create agent in the active workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentCreate" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/agents/{id}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    get:
      tags: [Agents]
      operationId: getAgent
      summary: Read agent (api_key masked)
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Agents]
      operationId: patchAgent
      summary: Update agent (omit `api_key` to preserve it)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentPatch" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Agents]
      operationId: deleteAgent
      summary: Delete agent (409 if any chat references it)
      responses:
        "200": { $ref: "#/components/responses/Removed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/agents/{id}/probe:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    post:
      tags: [Agents]
      operationId: probeAgent
      summary: One-shot prompt against the agent (does not persist)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: "User prompt. Defaults to `Hello` when missing or empty."
                  default: Hello
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [content]
                properties:
                  content: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
...
  /v1/media:
    post:
      tags: [Media]
      operationId: uploadMedia
      summary: Upload binary
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type, file]
              properties:
                type:
                  type: string
                  enum: [image, audio, video, document, sticker]
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: uploaded
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: string, format: uuid }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
  /v1/media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      tags: [Media]
      operationId: getMedia
      summary: Read metadata + download URL
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/MediaRecord"
                  - type: object
                    required: [download_url]
                    properties:
                      download_url: { type: string, format: uri }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Media]
      operationId: deleteMedia
      summary: Delete media
      responses:
        "200": { $ref: "#/components/responses/Removed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/media/{mediaId}/download:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      tags: [Media]
      operationId: downloadMedia
      summary: Stream binary content
      responses:
        "200":
          description: ok
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }